requirements.txt
osmnx>=1.9.0
networkx>=3.2
neat-python>=0.92
numpy>=1.26
pandas>=2.1
matplotlib>=3.8
scikit-learn>=1.3
